pyproject.toml
[tool.poetry]
name = "vesci"
version = "0.0.0"
description = "Virtual Environment and Source Control Integration for Python"
readme = "README.md"
authors = ["David Naughton <naughton@umn.edu>"]
license = "MIT"

#[tool.poetry.scripts]
#ve = 'vesci.vesci:cli'

[tool.poetry.dependencies]
python = "^3.10.0"
addict = "*"
requests = "^2.27.1"
GitPython = "^3.1.24"
click = "^8.0.3"

[tool.poetry.dev-dependencies]
pytest = "*"
python_dotenv = "*"
pytest-order = "^1.0.0"
